# docs/openapi.yaml
# Здесь будет OpenAPI описание всех доступных эндпоинтов и моделей данных.
openapi: 3.0.3

info:
  title: Bank Cards Management API
  description: |
    REST API для управления банковскими картами.
    Аутентификация через JWT, роли ADMIN и USER.
  version: 1.0.0

servers:
  - url: http://localhost:8080

security:
  - bearerAuth: []

tags:
  - name: Auth
  - name: Users
  - name: Cards
  - name: Transfers

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: user1
        password:
          type: string
          example: password123

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: user1
        password:
          type: string
          example: password123

    CreateCardRequest:
      type: object
      required: [userId, cardNumber, expirationDate, initialBalance]
      properties:
        userId:
          type: integer
          example: 1
        cardNumber:
          type: string
          example: "4111111111111111"
        expirationDate:
          type: string
          format: date
          example: "2028-12-31"
        initialBalance:
          type: number
          format: decimal
          example: 1000.00

    CardResponse:
      type: object
      properties:
        id:
          type: integer
          example: 10
        maskedNumber:
          type: string
          example: "**** **** **** 1234"
        owner:
          type: string
          example: user1
        expiryDate:
          type: string
          format: date
        status:
          type: string
          enum: [ACTIVE, BLOCKED, BLOCK_REQUESTED, EXPIRED]
        balance:
          type: number
          format: decimal
          example: 500.00

    TransferRequest:
      type: object
      required: [fromCardId, toCardId, amount]
      properties:
        fromCardId:
          type: integer
          example: 1
        toCardId:
          type: integer
          example: 2
        amount:
          type: number
          format: decimal
          example: 100.00

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

paths:

  /api/users/register:
    post:
      tags: [Auth]
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        200:
          description: Пользователь зарегистрирован

  /api/users/login:
    post:
      tags: [Auth]
      summary: Логин и получение JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: JWT token

  /api/users/me:
    get:
      tags: [Users]
      summary: Получить данные текущего пользователя
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK

  /api/users/getAllUsers:
    get:
      tags: [Users]
      summary: Получить всех пользователей (ADMIN)
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
        403:
          description: Forbidden

  /api/cards/createCard:
    post:
      tags: [Cards]
      summary: Создать карту (ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        200:
          description: Card created

  /api/cards/getAllCards:
    get:
      tags: [Cards]
      summary: Получить все карты (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: size
          schema:
            type: integer
      responses:
        200:
          description: OK

  /api/cards/my:
    get:
      tags: [Cards]
      summary: Получить свои карты (USER)
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK

  /api/cards/{id}/balance:
    get:
      tags: [Cards]
      summary: Получить баланс карты (USER)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK

  /api/cards/{id}/request-block:
    get:
      tags: [Cards]
      summary: Запросить блокировку карты (USER)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK

  /api/cards/block/{id}:
    put:
      tags: [Cards]
      summary: Заблокировать карту (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK

  /api/cards/activate/{id}:
    put:
      tags: [Cards]
      summary: Активировать карту (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK

  /api/cards/transfer:
    post:
      tags: [Transfers]
      summary: Перевод между своими картами (USER)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        200:
          description: Transfer completed
        400:
          description: Invalid request